+++ __tests__/transformers/callouts.test.ts
@@ -125,6 +125,54 @@ describe('callouts transformer', () => {
     `);
   });
 
+  it('can parse callouts with inline code in the heading', () => {
+    const md = `
+> 🚧 \`It works!\`
+>
+> And, it no longer deletes your content!
+`;
+    const tree = mdast(md);
+    removePosition(tree, { force: true });
+
+    expect(tree.children[0]).toMatchInlineSnapshot(`
+      {
+        "children": [
+          {
+            "children": [
+              {
+                "type": "text",
+                "value": "",
+              },
+              {
+                "type": "inlineCode",
+                "value": "It works!",
+              },
+            ],
+            "depth": 3,
+            "type": "heading",
+          },
+          {
+            "children": [
+              {
+                "type": "text",
+                "value": "And, it no longer deletes your content!",
+              },
+            ],
+            "type": "paragraph",
+          },
+        ],
+        "data": {
+          "hName": "Callout",
+          "hProperties": {
+            "icon": "🚧",
+            "theme": "warn",
+          },
+        },
+        "type": "rdme-callout",
+      }
+    `);
+  });
+
   it('can parse a jsx callout into a rdme-callout', () => {
     const md = `
 <Callout icon="📘" theme="info">
@@ -161,5 +209,38 @@ describe('callouts transformer', () => {
     const tree = mdast(md);
 
     expect(tree.children[0].data.hProperties).toHaveProperty('theme', 'info');
+  });
+
+  it('can correctly wrap a heading around a callout with a complex title and preserve the correct position data', () => {
+    const md = '> 📘 This is a callout [**with** a _link_](https://example.com)';
+
+    const tree = mdast(md);
+
+    // @ts-expect-error -- children should be defined
+    expect(tree.children[0].children[0].position).toMatchInlineSnapshot(`
+      {
+        "end": {
+          "column": 64,
+          "line": 1,
+          "offset": 63,
+        },
+        "start": {
+          "column": 6,
+          "line": 1,
+          "offset": 5,
+        },
+      }
+    `);
+  });
+
+  it('can parse a jsx callout and set a theme from the icon "👍"', () => {
+    const md = `
+<Callout icon="👍">
+### This is a callout
+</Callout>`;
+
+    const tree = mdast(md);
+
+    expect(tree.children[0].data.hProperties).toHaveProperty('theme', 'okay');
   });
 });